Move back to u16 stack size, allocate full 0-255 register range
Also simplify some things.  This is in preparation for the future, where
functions will be given the full 0-255 range no matter what, and this should
speed up stack access for most instructions

+++ src/function.rs
@@ -23,7 +23,7 @@ pub enum UpValueDescriptor {
 pub struct FunctionProto<'gc> {
     pub fixed_params: u8,
     pub has_varargs: bool,
-    pub stack_size: u8,
+    pub stack_size: u16,
     pub constants: Vec<Value<'gc>>,
     pub opcodes: Vec<OpCode>,
     pub upvalues: Vec<UpValueDescriptor>,
